DRY iterators (moved to DotMatrix)
