+++ src/data/CharDot.java
@@ -82,85 +82,7 @@ public class CharDot
 
     public Iterator<Dot[]> iterator(ScrollDirection direction)
     {
-        if (direction == ScrollDirection.LEFT)
-        {
-            return new Iterator<>()
-            {
-                int index = 0;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index < hLength && charDots.getCol(index) != null;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return charDots.getCol(index++);
-                }
-            };
-        }
-
-        else if (direction == ScrollDirection.RIGHT)
-        {
-            return new Iterator<>()
-            {
-                int index = hLength - 1;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index >= 0 && charDots.getCol(index) != null;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return charDots.getCol(index--);
-                }
-            };
-        }
-
-        else if (direction == ScrollDirection.UP)
-        {
-            return new Iterator<>()
-            {
-                int index = 0;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index < vLength && charDots.getRow(index) != null;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return charDots.getRow(index++);
-                }
-            };
-        }
-
-        else // DOWN
-        {
-            return new Iterator<>()
-            {
-                int index = vLength - 1;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index >= 0 && charDots.getRow(index) != null;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return charDots.getRow(index--);
-                }
-            };
-        }
+        return charDots.iterator(direction);
     }
 
     public static void initMap()
