+++ src/data/ImageSegment.java
@@ -40,84 +40,6 @@ public abstract class ImageSegment extends Segment
     @Override
     public Iterator<Dot[]> iterator(ScrollDirection direction)
     {
-        if (direction == ScrollDirection.LEFT)
-        {
-            return new Iterator<>()
-            {
-                int index = 0;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index < hlength;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return contents.getCol(index++);
-                }
-            };
-        }
-
-        else if (direction == ScrollDirection.RIGHT)
-        {
-            return new Iterator<>()
-            {
-                int index = hlength - 1;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index >= 0;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return contents.getCol(index--);
-                }
-            };
-        }
-
-        else if (direction == ScrollDirection.UP)
-        {
-            return new Iterator<>()
-            {
-                int index = 0;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index < vlength;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return contents.getRow(index++);
-                }
-            };
-        }
-
-        else // DOWN
-        {
-            return new Iterator<>()
-            {
-                int index = vlength - 1;
-
-                @Override
-                public boolean hasNext()
-                {
-                    return index >= 0;
-                }
-
-                @Override
-                public Dot[] next()
-                {
-                    return contents.getRow(index--);
-                }
-            };
-        }
+        return contents.iterator(direction);
     }
 }
